Update all feature vectors to actual browser data
For some unknown reason everything that was prepared offline with gv-detect did not work live, so all vectors are updated to match what is reported by browser.

Vectors are only reported in debug mode (with query argument debug=1)

Also a new UI mode "show all marks" is added, because it looks like extra information is both sometimes imprecise and clutters the view.